convert PageStore and RootStore to typescript
